Reorganise/rewrite visualization files to work with new setup of dataloader/generators, normalise packages
